LayoutTests/http/tests/w3c/html/semantics/forms/the-input-element/datetime-2-expected.txt


PASS empty value 
PASS datetime input value set to 2014-01-01T11:11Z without min/max 
FAIL datetime input value set to 2014-01-01 11:11Z without min/max assert_equals: expected "2014-01-01T11:11Z" but got "2014-01-01 11:11Z"
FAIL datetime input value set to 2014-01-01 11:11-04:00 without min/max assert_equals: expected "2014-01-01T15:11Z" but got "2014-01-01 11:11-04:00"
FAIL datetime input value set to 2014-01-01 11:11-0400 without min/max assert_equals: expected "2014-01-01T15:11Z" but got "2014-01-01 11:11-0400"
FAIL datetime input value set to 2014-01-01 11:11:00.000-04:00 without min/max assert_equals: expected "2014-01-01T15:11Z" but got "2014-01-01 11:11:00.000-04:00"
FAIL datetime input value set to 2014-01-01  11:11:00.000Z without min/max assert_equals: expected "" but got "2014-01-01  11:11:00.000Z"
FAIL datetime input value set to 2014-01-01 11:11:00.000 without min/max assert_equals: expected "" but got "2014-01-01 11:11:00.000"
FAIL datetime input value set to 2014-01-01 11:11:00.000+ without min/max assert_equals: expected "" but got "2014-01-01 11:11:00.000+"
FAIL datetime input value set to 2014-01-01 11:11:00.000+24 without min/max assert_equals: expected "" but got "2014-01-01 11:11:00.000+24"
FAIL datetime input value set to 2014-01-01 11:11:00.000+2360 without min/max assert_equals: expected "" but got "2014-01-01 11:11:00.000+2360"
FAIL datetime input value set to 2014-01-0 11:11:00.000+0400 without min/max assert_equals: expected "" but got "2014-01-0 11:11:00.000+0400"
FAIL datetime input value set to 2014-01-01 11:1:00.000+0400 without min/max assert_equals: expected "" but got "2014-01-01 11:1:00.000+0400"
FAIL Value >= min attribute assert_equals: expected "2014-01-01T11:11Z" but got "2014-01-01 11:11Z"
FAIL Value < min attribute assert_equals: expected "2014-01-01T11:11Z" but got "2014-01-01 11:10Z"
FAIL Value <= max attribute assert_equals: expected "2014-01-01T11:11Z" but got "2014-01-01 11:10Z"
FAIL Value > max attribute assert_equals: expected "2014-01-01T11:10Z" but got "2014-01-01 11:11Z"
